+++ src/main/java/com/htmake/htbot/discord/commands/battle/util/BattleUtil.java
@@ -2,7 +2,7 @@ package com.htmake.htbot.discord.commands.battle.util;
 
 import com.htmake.htbot.discord.skillAction.condition.Condition;
 import com.htmake.htbot.discord.commands.dungeon.data.GetItem;
-import com.htmake.htbot.discord.skillAction.condition.extend.Faint;
+import com.htmake.htbot.discord.skillAction.condition.extend.etc.Faint;
 import com.htmake.htbot.discord.util.FormatUtil;
 import com.htmake.htbot.discord.util.MessageUtil;
 import com.htmake.htbot.discord.util.ObjectMapperUtil;
