+++ src/main/java/com/htmake/htbot/discord/skillAction/skills/warrior/class3/S1304.java
@@ -6,7 +6,7 @@ import com.htmake.htbot.discord.commands.battle.data.status.extend.MonsterStatus
 import com.htmake.htbot.discord.commands.battle.data.status.extend.PlayerOriginalStatus;
 import com.htmake.htbot.discord.commands.battle.data.status.extend.PlayerStatus;
 import com.htmake.htbot.discord.skillAction.condition.Condition;
-import com.htmake.htbot.discord.skillAction.condition.extend.SwordGod;
+import com.htmake.htbot.discord.skillAction.condition.extend.etc.SwordGod;
 import com.htmake.htbot.discord.skillAction.skill.impl.AbstractSkillStrategy;
 import com.htmake.htbot.discord.skillAction.type.SkillType;
 import kotlin.Pair;
