+++ src/main/java/com/htmake/htbot/discord/skillAction/skills/wizard/class1/S3103.java
@@ -5,7 +5,7 @@ import com.htmake.htbot.discord.commands.battle.data.PlayerData;
 import com.htmake.htbot.discord.commands.battle.data.status.extend.MonsterStatus;
 import com.htmake.htbot.discord.commands.battle.data.status.extend.PlayerStatus;
 import com.htmake.htbot.discord.skillAction.condition.Condition;
-import com.htmake.htbot.discord.skillAction.condition.extend.Faint;
+import com.htmake.htbot.discord.skillAction.condition.extend.etc.Faint;
 import com.htmake.htbot.discord.skillAction.skill.action.SkillAction;
 import com.htmake.htbot.discord.skillAction.skill.impl.AbstractSkillStrategy;
 import com.htmake.htbot.discord.skillAction.type.SkillType;
