+++ src/main/java/com/htmake/htbot/discord/skillAction/skills/wizard/class3/S3305.java
@@ -4,7 +4,7 @@ import com.htmake.htbot.discord.commands.battle.data.MonsterData;
 import com.htmake.htbot.discord.commands.battle.data.PlayerData;
 import com.htmake.htbot.discord.commands.battle.data.status.extend.PlayerStatus;
 import com.htmake.htbot.discord.skillAction.condition.Condition;
-import com.htmake.htbot.discord.skillAction.condition.extend.Realize;
+import com.htmake.htbot.discord.skillAction.condition.extend.etc.Realize;
 import com.htmake.htbot.discord.skillAction.skill.action.SkillAction;
 import com.htmake.htbot.discord.skillAction.skill.impl.AbstractSkillStrategy;
 import com.htmake.htbot.discord.skillAction.type.SkillType;
